basic xDAI wallet is done
