+++ www/index.haml
@@ -37,4 +37,7 @@
     -# %script{ src: "node_modules/hyperhtml/min.js",  type: "text/javascript" }
     -# %script{ src: "node_modules/web3/dist/web3.umd.js",  type: "text/javascript" }
 
+    %script{ src: "js/utils.js", type: "text/javascript" }
+    %script{ src: "js/view.js", type: "text/javascript" }
+    %script{ src: "js/app.js", type: "text/javascript" }
     %script{ src: "js/index.js", type: "text/javascript" }
