+++ www/js/utils.js
@@ -1,0 +1,6 @@
+const doc = document
+
+class NullKeychain {}
+class NullBalance {}
+class NullRate {}
+class NullEventEmitter {}
